Trading Bot/requirements.txt
python-binance==1.0.35
requests==2.31.0
argparse==1.4.0